+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - The mod has been completely remade, making it easier to program and reducing the amount of errors.
 - More ways to send messages. Examples are in the `/cobbleutils sendMessage` command.
 - Added formula, this is more flexible than filters.
+- Added trade notifications, which can be customized in the config.
 
 ### Bug Fixes